+++ packages/behaviors/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @refrakt-md/behaviors
 
+## 0.9.8
+
 ## 0.9.7
 
 ## 0.9.6