+++ packages/highlight/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @refrakt-md/highlight
 
+## 0.9.8
+
+### Patch Changes
+
+- @refrakt-md/transform@0.9.8
+- @refrakt-md/types@0.9.8
+
 ## 0.9.7
 
 ### Patch Changes