+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @refrakt-md/svelte
 
+## 0.9.8
+
+### Patch Changes
+
+- @refrakt-md/behaviors@0.9.8
+- @refrakt-md/transform@0.9.8
+- @refrakt-md/types@0.9.8
+
 ## 0.9.7
 
 ### Patch Changes