+++ packages/sveltekit/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @refrakt-md/sveltekit
 
+## 0.9.8
+
+### Patch Changes
+
+- @refrakt-md/types@0.9.8
+
 ## 0.9.7
 
 ### Patch Changes